requirements.txt
typer[all]==0.9.0
pymupdf==1.23.8
faiss-cpu==1.7.4
sentence-transformers>=3.0.0
langchain==0.0.350
pydantic>=2.7.0
duckdb==0.9.2
pandas==2.2.0
jinja2==3.1.2
rich==13.7.0
beautifulsoup4==4.12.2
